Moved SubmissionResult to message module. Made SubmissionResult frozen. Added again poetry.lock file.
